# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A lightweight real-time chat backend written in Go. The server manages ephemeral chat rooms where clients connect via WebSocket to exchange messages. Rooms are created on-demand and automatically deleted after 3 hours of inactivity.\n\n## Additional Info\nMany resources (rooms, users, messages) carry an `additionalInfo` field. This is a free-form JSON object that the server stores as-is — it has no predefined schema and is never validated or interpreted by the backend.\n\nUse it to attach arbitrary metadata to any resource, for example:\n- **Rooms:** theme, description, language, or feature flags for your UI.\n- **Users:** avatar URL, bio, status text, or client-specific preferences.\n- **Messages:** formatting hints, link previews, or custom reaction data.\n\nThe field is always optional. If omitted it defaults to an empty object (`{}`). On `PATCH` requests the provided keys are merged into the existing object; on `PUT` the entire object is replaced.",
        "title": "Chat Room API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/join/{roomID}": {
            "get": {
                "description": "Upgrades the HTTP connection to WebSocket and joins the requested room.\n\n**Authentication options:**\n- `userId` (UUID): Join as a registered user from the registry. Takes precedence over `user`.\n- `user` (string): Join as an ephemeral user with the given display name.\n- Neither: Server assigns a random display name.\n\n**User info extraction:** Set `userInfo=true` to receive a self-join message with a `self` flag, allowing clients to extract their user information.\n\n**Connection management:** Server sends ping every 30s, expects pong within 60s. Max message size: 10 MiB.",
                "tags": [
                    "websocket"
                ],
                "summary": "Join room via WebSocket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Room ID",
                        "name": "roomID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registered user UUID",
                        "name": "userId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ephemeral display name",
                        "name": "user",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Enable self-join message with user info",
                        "name": "userInfo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols - WebSocket connection established"
                    },
                    "400": {
                        "description": "invalid room or user ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "room or user not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}